DEV: restructure tests: fix static paths
# Conflicts:
#	tests/TestHelper.php
